requirements.txt
django==4.2.2
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2==2.9.6
djangorestframework-simplejwt==5.2.2